zebra, bgpd: reevaluate dependant BGP routes whenever an SR-policy is updated (WIP)
This is a brute-force workaround solution that reevaluates *all*
BGP routes whose nexthop is the same as the endpoint of the SR
policy that was updated. Ideally only the affected colored BGP
routes should be reevaluated.

Also, bgpd is using the same nexthop cache for both colored and
non-colored routes, which can lead to incorrect behavior in some
specific circumstances (especially for labeled BGP routes). This
and other issues should be addressed in a subsequent commit.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>