ospf6d: implement Intra Prefix LA bit correctly
For loopback, PtP and PtMP, the correct thing is to include only our
local addresses as a /128 with the LA bit set (the prefix length being
the important part, LA is more or less tangential.)  For PtP/PtMP, this
ensures that we don't accidentally cover another router's addresses on a
broken link/in a partitioned network.  For loopback the addresses should
be configured as /128 anyway so it's just footgun prevention.

Signed-off-by: David Lamparter <equinox@opensourcerouting.org>