tests: stop using fifos in bgp...multipath_relax
While this works correctly if everything is set up correctly, it fails
very badly if anything is off.  Particularly, writing into a fifo that
has no reader on it will hang indefinitely, making it quite annoying to
debug.

Open a proper unix pipe and pass it down instead.  Will still fail in
the same situation, but it will fail in a way that is much easier to
understand and address.

(Brought to you by hours of debugging that could've been spent better
elsewhere :/)

Signed-off-by: David Lamparter <equinox@opensourcerouting.org>